原来count采用int，而add函数的参数为size_t，在不同平台可能会不兼容。 全部统一成unsigned int

+++ src/chaos/shader/ChsShaderUniform.cpp
@@ -69,7 +69,7 @@ namespace Chaos {
 	}
 	
 	//----------------------------------------------------------------------------------------------
-	void ChsShaderUniform::add( std::string name, void * varAddr, ChsShaderUniformDataType type, size_t count ){
+	void ChsShaderUniform::add( std::string name, void * varAddr, ChsShaderUniformDataType type, unsigned int count ){
 		if( uniformVariables.find( name ) != uniformVariables.end() )
 			return;//already in list, do nothing
 		//leave location with -1
